% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseMeanBar.R
\name{baseMeanBar}
\alias{baseMeanBar}
\title{baseMeanBar: Barplot of your genes of interest!}
\usage{
baseMeanBar(deseqAbs, genes, cond = NULL)
}
\arguments{
\item{deseqAbs:}{deseqAbs object with baseMean and test data}

\item{genes:}{a vector of gene IDs to be extracted from rownames of data}

\item{cond:}{a vector with condition names to use (must match column names of baseMean df)}
}
\description{
barplot of gene expression
}
\examples{
deseqAbs <- dnmt
genes <- c("DNMT1","TRIM28","PAX6","DCX","SOX2","AGO2")
baseMeanBar(dnmt,genes)
}
