% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeGenes.R
\name{closeGenes}
\alias{closeGenes}
\title{closeGenes : Get close features (genes )}
\usage{
closeGenes(a, b, dist = 10000)
}
\arguments{
\item{a:}{a .bed position of all genes of interest. Must have $Chr, $Start, $End $ID $Strand column names! TSS of the transcrips will be used}

\item{b:}{a .bed position of all features of interest. Must have $Chr, $Start, $End $ID $Strand column names! TSS of the features will be used}

\item{dist:}{how many basepairs from features B should a feature A be to be included? Default = 10000bps}
}
\description{
Get genes (or other features) close to some other features (.bed). TSS used.
}
\examples{
closeGenes <- enesClose(a = gencode.transcripts, b = L1s.up, distance = 50000 )
}
