% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesClose.R
\name{genesClose}
\alias{genesClose}
\title{genesClose : Get genes close to features}
\usage{
genesClose(genPos, featPos, dist)
}
\arguments{
\item{pos:}{position of all genes of interest (e.g. deseqAbsObj$pos). Must have $Chr, $Start, $End columns and gene names as rownames. If the annotation is collapsed (such as after featureCounts), only on tss (the most extreme) is used for each gene.}

\item{feat:}{a .bed file with position that genes should be close to (bed: chr - start - end - ID - . - strand)}

\item{distance:}{how far from features should a gene be to be included?}
}
\description{
Plots scatter plot with log2FC on log2(baseMean), and marks significant genes
}
\examples{
closeGenes <- enesClose(pos = desqAobj$pos, feat=L1.up.bed, distance = 10000 )
}
