% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseMean.R
\name{getAverage}
\alias{getAverage}
\title{Get average read numbers within each condition}
\usage{
getAverage(dds)
}
\arguments{
\item{dds:}{dds object. MUST have condition as colData grouping!}
}
\description{
Calculates average expression of each group, based on condition of dds object. Returns list with Mean and SD values for each gene in each condition
}
\examples{
dds <- DESeqDataSetFromMatrix(countData = countData, colData = colData, desing =~ condition)
baseMeans <- getAverage(dds)
means <- baseMeans$Mean
SD <- baseMeans$SD
}
