% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubID.R
\name{getSubID}
\alias{getSubID}
\title{getSubID - easy parsing}
\usage{
getSubID(uniqID = NULL)
}
\arguments{
\item{uniqID:}{vector with unique IDs}
}
\description{
takes a vector of RepeatMasker TE_IDs with unique _x tag, and removes the tag.
}
\examples{
getSubID(uniqID = rownames(countTEdata))
}
