% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_geneFC.R
\name{GO_geneFC}
\alias{GO_geneFC}
\title{GOanalysis in R - GO terms}
\usage{
GO_geneFC(dabs = NULL, species = "hsa", BP = T, MF = F, CC = F,
  sameDir = T)
}
\arguments{
\item{dds:}{deseq object}

\item{species:}{only human (hsa) annotation currently supported, mm10 coming soon..}

\item{BP:}{test for biological process terms}

\item{MF:}{test for molecular function terms}

\item{CC:}{test for cellular compartment terms}

\item{sameDir:}{test one-directional, terms enriched with genes changed in only same direction}
}
\description{
Do GO term enrichement and analysis for a list of genes
}
\examples{
dabs <- deseqAbs$new(name="drugTest",colData=colDat,file=pathToFeatureCountsOutput)
GO_pathview(dabs = dabs,species = "hsa",BP=T,MF=F,CC=F,sameDir=T)
}
