% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_pathview.R
\name{GO_pathview}
\alias{GO_pathview}
\title{GOanalysis in R  - pathways}
\usage{
GO_pathview(dabs = NULL, species = "hsa", Npathways = 5)
}
\arguments{
\item{dds:}{deseq object}

\item{species:}{only human (hsa) annotation currently supported, mm10 coming soon}

\item{Npathways:}{How many pathways to plot}
}
\description{
Do GO term enrichement and analysis for a list of genes (pathway analysis and plots)
}
\examples{
dabs <- deseqAbs$new(name="drugTest",colData=colDat,file=pathToFeatureCountsOutput)
GO_pathview(dabs = dabs,species = "hsa",Npathways = 4)
}
