% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_topGO_geneSet.R
\name{GO_topGO_geneSet}
\alias{GO_topGO_geneSet}
\title{GOanalysis in R - topGO enrichment test terms}
\usage{
GO_topGO_geneSet(dabs = NULL, geneSet = NULL, org = "hsa",
  term = "BP", nodeSize = 5, outdir = ".")
}
\arguments{
\item{dds:}{deseq object}

\item{org:}{only hg38 annotation currently supported, mm10 coming soon..}

\item{term:}{which go term ("BP" "MF" or "CC")}

\item{geneSet:}{list of genes for test}

\item{nodeSize:}{Set smallest included node size (number of genes in term) in enrichment test}

\item{outdir:}{name of output directory}
}
\description{
Do GO term enrichement and analysis for a list of genes
}
\examples{
dabs <- deseqAbs$new(name="drugTest",colData=colDat)
dabs$makeDiffex
geneSet <- getSignName(x = dabs$test$Default,p=0.01)$up # get upregulated genes
GO_topGO_geneSet(dabs = dabs,org = "hsa",BP=T,MF=F,CC=F,geneSet=geneSet,outdir=currdir,nodeSize=3)
}
