% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_topGO_rank.R
\name{GO_topGO_rank}
\alias{GO_topGO_rank}
\title{GOanalysis in R - topGO enrichment test terms RANK based}
\usage{
GO_topGO_rank(dabs = NULL, org = "hsa", term = "BP", nodeSize = 5,
  rank = "log2fc", sigCut = 1)
}
\arguments{
\item{dds:}{deseq object}

\item{org:}{only hg38 annotation currently supported, mm10 coming soon..}

\item{term:}{which go term ("BP" "MF" or "CC")}

\item{nodeSize:}{Set smallest included node size (number of genes in term) in enrichment test}

\item{rank:}{How to rank genes (either "log2fc" or "padj")}

\item{sigCut:}{where to cut for sign genes. when rank="log2fc", it is the absolute log2FC. when rank="padj", it is p-adj value..}
}
\description{
Do GO term enrichement and analysis for a list of genes
}
\examples{
dabs <- deseqAbs$new(name="drugTest",colData=colDat,file=pathToFeatureCountsOutput)
dabs$makeDiffex
geneSet <- getSignName(x = dabs$test$Default,p=0.01)$up # get upregulated genes
GO_topGO_geneSet_rank(dabs = dabs,org = "hsa",BP=T,MF=F,CC=F,geneSet=geneSet)
}
