% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatGenes.R
\name{heatGenes}
\alias{heatGenes}
\title{HeatGenes: Heatmap of your genes of interest!}
\usage{
heatGenes(data, genes, a1 = NULL, a2 = NULL, n1 = NULL, n2 = NULL, sd)
}
\arguments{
\item{data:}{deseq matrix data (with varianceStablizing, rlog etc - eg. assay(vsd)}

\item{genes:}{a vector of gene IDs that can be grepped from rownames of data}

\item{a1:}{annotation of the samples}

\item{a2:}{annotation of the samples}

\item{n1:}{name of annotation in a1}

\item{n2:}{name of annotation in a2}

\item{sd:}{standard deviation cutoff for heatmap}
}
\description{
Plots heatmap showing the expression of the genes of given gene-list and having sd above given value
}
