% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanBarGeneric.R
\name{meanBarGeneric}
\alias{meanBarGeneric}
\title{meanBarGeneric: Barplot of your genes of interest!}
\usage{
meanBarGeneric(tab, cond = NULL, points = FALSE, jitter = F,
  ylab = "")
}
\arguments{
\item{tab:}{data.frame to plot!}

\item{cond:}{a vector of conditions to group data}

\item{points:}{Set to TRUE if points should be added}

\item{jitter:}{set to TRUE if jitter should be applied to points}

\item{ylab:}{enter ylab for plot!}
}
\description{
barplot of gene expression
}
