% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significantHeat.R
\name{mostSignificantHeat}
\alias{mostSignificantHeat}
\title{Plot most significant genes: heatmap}
\usage{
mostSignificantHeat(data, test, ntop = 50, a1 = NULL, a2 = NULL,
  n1 = NULL, n2 = NULL)
}
\arguments{
\item{data:}{normalized deseq matrix data (with varianceStablizing, rlog etc - eg. assay(vsd)}

\item{test:}{output from results(dds)}

\item{ntop:}{how many genes to display (the <ntop> most variable genes)}

\item{a1:}{annotation of the samples}

\item{a2:}{annotation of the samples}

\item{n1:}{name of annotation in a1}

\item{n2:}{name of annotation in a2}
}
\description{
Plots heatmap showing the expression of most variable genes
}
\examples{
vst <- varianceStabilizingTransformation(dds)
test <- results(dds)
mostSignificantHeat(data = assay(vst),test=test,ntop = 100, a1 = colData$cellLine, a2 = colData$treatment,n1="Cell Line",n2="Treatment")
}
