% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAplot.R
\name{PCAplotter}
\alias{PCAplotter}
\title{PCAplotter my awesome function #1}
\usage{
PCAplotter(dat, ntop = 5000, color = NULL, shape = NULL, title = "",
  label = NULL)
}
\arguments{
\item{dat:}{normalized deseq data (with varianceStablizing, rlog etc)}

\item{ntop:}{how many genes to include}

\item{color:}{how to group colors}

\item{shape:}{how to group shapes. (by default NULL if only color by groups wanted)}

\item{title:}{title of plot}

\item{label:}{vector of labels for each point (optional)}
}
\description{
creates PCA plot
}
\examples{
vst <- varianceStabilizingTransformation(dds);
PCAplotter(dat = vst,ntop = 1000,color = colData$cellType, shape = colData$treatment,title = "PCA plot top1000 genes", label = colData$sample)
}
