% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanBar.R
\name{rpkmMeanBar}
\alias{rpkmMeanBar}
\alias{meanBar}
\title{rpkmMeanBar: Barplot of your genes of interest!}
\usage{
meanBar(deseqAbs, genes, cond = NULL, rpkm = FALSE)
}
\arguments{
\item{deseqAbs:}{deseqAbs object with rpkmMean and test data}

\item{genes:}{a vector of gene IDs to be extracted from rownames of data}

\item{cond:}{a vector of conditions to plot}

\item{rpkm:}{Set to TRUE if RPKM should be plotted instead of baseMean}
}
\description{
barplot of gene expression
}
\examples{
deseqAbs <- dnmt
genes <- c("DNMT1","TRIM28","PAX6","DCX","SOX2","AGO2")
rpkmMeanBar(dnmt,genes)
}
