% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseMeanBar.R
\name{baseMeanBar}
\alias{baseMeanBar}
\title{baseMeanBar: Barplot of your genes of interest!}
\usage{
baseMeanBar(deseqAbs, genes)
}
\arguments{
\item{deseqAbs:}{deseqAbs object with baseMean and test data}

\item{genes:}{a vector of gene IDs to be extracted from rownames of data}
}
\description{
barplot of gene expression
}
