% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenes.R
\name{getGenes}
\alias{getGenes}
\title{getGenes: Get data for your genes of interest!}
\usage{
getGenes(data, genes)
}
\arguments{
\item{data:}{matrix data (can be any format, but MUST have geneID as rownames)}

\item{genes:}{a vector of gene IDs that can be grepped from rownames of data}
}
\description{
gets data for given genes. data can be of any format, as long as the genes IDs are rownames
}
