% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSign.R
\name{getSign}
\alias{getSign}
\title{getSign genes as table with test data}
\usage{
getSign(x, p = 0.01, l = 0.2)
}
\arguments{
\item{x:}{object of results(dds), or a data.frame with column named "padj" with adjusted p-values, and one named "log2FoldChange" with log2(fold changes)}

\item{p:}{p-adj cutoff to use}

\item{l:}{log2(fc) cutoff to use}
}
\description{
Get the significantly changed genes, returning deseq-table with up and down-regulated genes separately in a list.
}
\examples{
test <- results(dds,contrast = c("condition","genex-KO","WT"))
sign <- getSign(x = test, p = 0.01, l = 0.5)
up <- sign$up
down <- sign$down
}
