% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatGenes.R
\name{heatGenes}
\alias{heatGenes}
\title{HeatGenes: Heatmap of your genes of interest!}
\usage{
heatGenes(data, genes, a1 = NULL, a2 = NULL, n1 = NULL, n2 = NULL,
  sd = 1, z = FALSE, cluster_col = T)
}
\arguments{
\item{data:}{deseq matrix data (with varianceStablizing, rlog etc - eg. assay(vsd)}

\item{genes:}{a vector of gene IDs to be extracted from rownames of data}

\item{a1:}{annotation of the samples}

\item{a2:}{annotation of the samples}

\item{n1:}{name of annotation in a1}

\item{n2:}{name of annotation in a2}

\item{sd:}{standard deviation cutoff for heatmap}

\item{z:}{logical operator. if TRUE, normalize by row (z-score). if FALSE (default), then plot normalized expression values}

\item{cluster_col:}{logical operator: if TRUE (default), cluster columns, if FALSE do not cluster columns}
}
\description{
Plots heatmap showing the expression of the genes of given gene-list and having sd above given value
}
\examples{
vst <- varianceStabilizingTransformation(dds)
genes <- c("TRIM28","DNMT1","ZNF52")
heatGenes(data = assay(vst),genes = genes, sd = 1, a1 = colData$cellLine, a2 = colData$treatment,n1 = "Cell Line",n2 = "Treatment",z = T)
}
