% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanBar.R
\name{meanBar}
\alias{meanBar}
\title{meanBar: Barplot of your genes of interest!}
\usage{
meanBar(deseqAbs, genes, cond = NULL, FPKM = FALSE, points = FALSE)
}
\arguments{
\item{deseqAbs:}{deseqAbs object with FPKMMean and test data}

\item{genes:}{a vector of gene IDs to be extracted from rownames of data}

\item{cond:}{a vector of conditions to plot}

\item{FPKM:}{Set to TRUE if FPKM should be plotted instead of baseMean}

\item{points:}{Set to TRUE if points should be added}

\item{deseqAbs:}{deseqAbs object with FPKMMean and test data}

\item{genes:}{a vector of gene IDs to be extracted from rownames of data}

\item{cond:}{a vector of conditions to plot}

\item{FPKM:}{Set to TRUE if FPKM should be plotted instead of baseMean}

\item{points:}{Set to TRUE if points should be added}
}
\description{
barplot of gene expression

barplot of gene expression
}
\examples{
genes <- c("DNMT1","TRIM28","PAX6","DCX","SOX2","AGO2")
meanBar(dnmt,genes)
genes <- c("DNMT1","TRIM28","PAX6","DCX","SOX2","AGO2")
meanBar(dnmt,genes)
}
