% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableGenesHeat.R
\name{mostVariableHeat}
\alias{mostVariableHeat}
\title{mostVariableHeat - heatmap of variable genes}
\usage{
mostVariableHeat(data, ntop = 50, a1 = NULL, a2 = NULL)
}
\arguments{
\item{data:}{normalized deseq data (with varianceStablizing, rlog etc)}

\item{ntop:}{how many genes to display (the <ntop> most variable genes)}

\item{a1:}{annotation of the samples}

\item{a2:}{annotation of the samples}
}
\description{
Plots heatmap showing the expression of most variable genes
}
