% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAplot.R
\name{PCAplotter}
\alias{PCAplotter}
\title{PCAplotter my awesome function #1}
\usage{
PCAplotter(dat, ntop, color, shape, title)
}
\arguments{
\item{dat:}{normalized deseq data (with varianceStablizing, rlog etc)}

\item{ntop:}{how many genes to include}

\item{color:}{how to group colors}

\item{shape:}{how to group shapes}

\item{title:}{title of plot}
}
\description{
creates PCA plot
}
