% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{volcano plot}
\usage{
volcanoPlot(test, max = NULL, p = 0.5, title = "transcriptome changes")
}
\arguments{
\item{test:}{output object from results() of DESeq2}

\item{p:}{cutoff of significance (p-adj)}

\item{title:}{string describing plot}
}
\description{
Generates volcano plot and marks significant genes. Scales the y axis if promted
}
